Properly serialize Mondrian Nodes when pickling
The Node struct has two array members, lower_bounds and upper_bounds.
Currently, __getstate__ just serializes the pointers to these arrays,
not the contents. When a Mondrian tree is pickled and unpickled in a
different address space, these pointers are dangling. Fixed by wrapping
the pointers in numpy arrays which pickle properly.

Fixes #55