Allow source install without build deps preinstall
This solves the issue where installing scikit-garden in an empty virtualenv does not work.

Adds numpy as setup requirements, and requires Cython only when running the build_ext, where Cython is needed. This pushes the Cython requirement to a point in the install where setuptools has had time to install Cython as a build requirement.

Replaces #77 which forgot to cythonize extensions.